service-python/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-multipart==0.0.9
pydantic==2.6.4
opencv-python==4.6.0.66
numpy==1.26.4
pytesseract==0.3.10
Pillow==10.2.0
paddleocr==2.7.3
paddlepaddle==2.6.2
pdf2image==1.17.0
